pyproject.toml
[tool.poetry]
name = "stock_collect"
version = "0.1.0"
description = "It is a service for me to collect my stock information from gmail."
authors = ["Eileen <lai.eling22@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.5"
google-api-python-client = "^2.8.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.4"
rich = "^10.3.0"
firebase-admin = "^5.0.1"
pandas = "^1.2.4"
xlrd = "^2.0.1"
openpyxl = "^3.0.7"
matplotlib = "^3.4.2"
lxml = "^4.6.3"
yfinance = "^0.1.63"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
isort = "^5.8.0"
autoflake = "^1.4"
mypy = "^0.902"
pytest = "^6.2.4"
pytest-mock = "^3.6.1"
mock = "^4.0.3"

[tool.poetry.scripts]
main = "stock_collect.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
